+++ CHANGELOG.md
@@ -25,7 +25,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Improve Search component. #91
   - Add event `input-change`.
   - Add prop `input-placeholder`.
-  - Add data `suggests`.
+  - Add data `suggestions`.
 
 ## [0.5.1][] - 2018-02-23