+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased][]
 
+## [0.5.2][] - 2018-03-05
+
 ### Added
 
 - Added the ["types" declaration file](https://github.com/archco/moss-ui/blob/master/index.d.ts). #92
@@ -154,7 +156,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Remove unnecessary part "aside.scss". #60
 - Remove deprecated scss mixin, `flex-container()`. instead use `flex-container-align`. #52
 
-[Unreleased]: https://github.com/archco/moss-ui/compare/v0.5.1...HEAD
+[Unreleased]: https://github.com/archco/moss-ui/compare/v0.5.2...HEAD
+[0.5.2]: https://github.com/archco/moss-ui/compare/v0.5.1...v0.5.2
 [0.5.1]: https://github.com/archco/moss-ui/compare/v0.5.0...v0.5.1
 [0.5.0]: https://github.com/archco/moss-ui/compare/v0.4.2...v0.5.0
 [0.4.2]: https://github.com/archco/moss-ui/compare/v0.4.1...v0.4.2